use anonymous producer and consumers

+++ src/dev/lpa/Order.java
@@ -1,4 +1,4 @@
 package dev.lpa;
 
-public record Order(int id, String shoeType, int quantity) {
+public record Order(long id, String shoeType, int quantity) {
 }
